requirements.txt
langchain
langchain-community
pypdf
gradio
chromadb
transformers
sentence-transformers
python-dotenv
langchain_groq
